+++ packages/core/src/api/exceptions/initialize.test.ts
@@ -1,11 +1,9 @@
 import { initializeFaro } from '../../initialize';
 import { mockConfig, mockInternalLogger, MockTransport } from '../../testUtils';
 import { TransportItem, TransportItemType } from '../../transports';
-import { dateNow } from '../../utils';
-import { mockMetas, mockTracesApi, mockTransports } from '../apiTestHelpers';
-import { USER_ACTION_CANCEL, USER_ACTION_END, USER_ACTION_START } from '../const';
-import { ItemBuffer } from '../ItemBuffer';
-import type { API, APIEvent, ApiMessageBusMessages } from '../types';
+import { mockMetas, mockTracesApi, mockTransports, mockUserActionsApi } from '../apiTestHelpers';
+import type { API } from '../types';
+import UserAction from '../userActions/userAction';
 
 import { initializeExceptionsAPI } from './initialize';
 import type { ExceptionEvent, ExceptionEventExtended, ExceptionStackFrame, PushErrorOptions } from './types';
@@ -233,17 +231,11 @@ describe('api.exceptions', () => {
       });
 
       it('Adds error cause to error context', () => {
-        // @ts-expect-error cause is missing in TS type Error
         const error = new Error('test', { cause: 'foo' });
-        // @ts-expect-error cause is missing in TS type Error
         const error2 = new Error('test2', { cause: [1, 3] });
-        // @ts-expect-error cause is missing in TS type Error
         const error3 = new Error('test3', { cause: { a: 'b' } });
-        // @ts-expect-error cause is missing in TS type Error
         const error4 = new Error('test4', { cause: new Error('original error') });
-        // @ts-expect-error cause is missing in TS type Error
         const error5 = new Error('test5', { cause: null });
-        // @ts-expect-error cause is missing in TS type Error
         const error6 = new Error('test6', { cause: undefined });
         const error7 = new Error('test6');
 
@@ -332,60 +324,32 @@ describe('api.exceptions', () => {
         expect((transport.items[0]?.payload as ExceptionEvent).value).toEqual(typeErrorMsg);
       });
     });
+  });
 
-    describe('User action', () => {
-      it('buffers the error if a user action is in progress', () => {
-        const internalLogger = mockInternalLogger;
-        const config = mockConfig();
+  describe('User action', () => {
+    it('buffers the error if a user action is in progress', () => {
+      const internalLogger = mockInternalLogger;
+      const config = mockConfig();
 
-        const actionBuffer = new ItemBuffer<TransportItem<APIEvent>>();
-
-        let message: ApiMessageBusMessages | undefined;
-
-        const getMessage = () => message;
-
-        message = {
-          type: USER_ACTION_START,
-          name: 'testAction',
-          startTime: Date.now(),
-          parentId: 'parent-id',
-        };
-        const api = initializeExceptionsAPI({
-          unpatchedConsole: console,
-          internalLogger,
-          config,
-          metas: mockMetas,
-          transports: mockTransports,
-          tracesApi: mockTracesApi,
-          actionBuffer,
-          getMessage,
-        });
-
-        api.pushError(new Error('test error'));
-        expect(actionBuffer.size()).toBe(1);
-
-        message = {
-          type: USER_ACTION_END,
-          name: 'testAction',
-          id: 'parent-id',
-          startTime: dateNow(),
-          endTime: dateNow(),
-          duration: 0,
-          eventType: 'click',
-        };
-
-        api.pushError(new Error('test error 2'));
-        expect(actionBuffer.size()).toBe(1);
-
-        message = {
-          type: USER_ACTION_CANCEL,
-          name: 'testAction',
-          parentId: 'parent-id',
-        };
-
-        api.pushError(new Error('test error 3'));
-        expect(actionBuffer.size()).toBe(1);
+      const api = initializeExceptionsAPI({
+        unpatchedConsole: console,
+        internalLogger,
+        config,
+        metas: mockMetas,
+        transports: mockTransports,
+        tracesApi: mockTracesApi,
+        userActionsApi: mockUserActionsApi,
       });
+
+      (mockUserActionsApi.getActiveUserAction as jest.Mock).mockReturnValueOnce(
+        new UserAction({
+          name: 'test',
+          trigger: 'foo',
+          transports: mockTransports,
+        })
+      );
+      api.pushError(new Error('test'));
+      expect(mockTransports.execute).not.toHaveBeenCalled();
     });
   });
 });