+++ packages/core/src/api/exceptions/initialize.ts
@@ -16,10 +16,8 @@ import {
   stringifyObjectValues,
 } from '../../utils';
 import { timestampToIsoString } from '../../utils/date';
-import { USER_ACTION_START } from '../const';
-import type { ItemBuffer } from '../ItemBuffer';
 import type { TracesAPI } from '../traces';
-import type { ApiMessageBusMessages } from '../types';
+import type { UserActionsAPI } from '../userActions';
 import { shouldIgnoreEvent } from '../utils';
 
 import { defaultExceptionType } from './const';
@@ -31,8 +29,7 @@ export function initializeExceptionsAPI({
   metas,
   transports,
   tracesApi,
-  actionBuffer,
-  getMessage,
+  userActionsApi,
 }: {
   unpatchedConsole: UnpatchedConsole;
   internalLogger: InternalLogger;
@@ -40,8 +37,7 @@ export function initializeExceptionsAPI({
   metas: Metas;
   transports: Transports;
   tracesApi: TracesAPI;
-  actionBuffer: ItemBuffer<TransportItem>;
-  getMessage: () => ApiMessageBusMessages | undefined;
+  userActionsApi: UserActionsAPI;
 }): ExceptionsAPI {
   internalLogger.debug('Initializing exceptions API');
 
@@ -107,9 +103,9 @@ export function initializeExceptionsAPI({
 
       internalLogger.debug('Pushing exception\n', item);
 
-      const msg = getMessage();
-      if (msg && msg.type === USER_ACTION_START) {
-        actionBuffer.addItem(item);
+      const activeUserAction = userActionsApi.getActiveUserAction();
+      if (activeUserAction) {
+        activeUserAction.addItem(item);
       } else {
         transports.execute(item);
       }