+++ packages/core/src/api/initialize.test.ts
@@ -1,11 +1,8 @@
-import { Observable, TransportItem } from '..';
 import { mockConfig, mockInternalLogger } from '../testUtils';
 import { unpatchedConsole } from '../unpatchedConsole/initialize';
 
 import { mockMetas, mockTransports } from './apiTestHelpers';
-import { apiMessageBus, initializeAPI } from './initialize';
-import { ItemBuffer } from './ItemBuffer';
-import * as createUserActionLifecycleHandlerModule from './userActionLifecycleHandler';
+import { initializeAPI } from './initialize';
 
 describe('initialize', () => {
   it('should initialize the API', () => {
@@ -31,23 +28,5 @@ describe('initialize', () => {
     expect(api).toHaveProperty('setUser');
     expect(api).toHaveProperty('setPage');
     expect(api).toHaveProperty('setView');
-  });
-
-  it('Exports the apiMessageBus observable', () => {
-    expect(apiMessageBus).toBeInstanceOf(Observable);
-  });
-
-  it('creates a user action lifecycle handler', () => {
-    const mockUserActionLifecycleHandler = jest.fn(() => ({
-      actionBuffer: new ItemBuffer<TransportItem>(),
-      getMessage: jest.fn(),
-    }));
-    jest
-      .spyOn(createUserActionLifecycleHandlerModule, 'createUserActionLifecycleHandler')
-      .mockImplementationOnce(mockUserActionLifecycleHandler);
-
-    initializeAPI(unpatchedConsole, mockInternalLogger, mockConfig(), mockMetas, mockTransports);
-
-    expect(mockUserActionLifecycleHandler).toHaveBeenCalled();
   });
 });