+++ packages/web-sdk/src/config/types.ts
@@ -1,4 +1,4 @@
-import type { Config, LogLevel } from '@grafana/faro-core';
+import type { Config } from '@grafana/faro-core';
 
 export interface BrowserConfig extends Partial<Omit<Config, 'app'>>, Pick<Config, 'app'> {
   url?: string;
@@ -7,7 +7,6 @@ export interface BrowserConfig extends Partial<Omit<Config, 'app'>>, Pick<Config
 
 export interface GetWebInstrumentationsOptions {
   captureConsole?: boolean;
-  captureConsoleDisabledLevels?: LogLevel[];
   enablePerformanceInstrumentation?: boolean;
   enableContentSecurityPolicyInstrumentation?: boolean;
 }