+++ packages/web-sdk/src/index.ts
@@ -17,7 +17,7 @@ export {
   CSPInstrumentation,
   UserActionInstrumentation,
 } from './instrumentations';
-export type { ConsoleInstrumentationOptions, ErrorEvent, ExtendedPromiseRejectionEvent } from './instrumentations';
+export type { ErrorEvent, ExtendedPromiseRejectionEvent } from './instrumentations';
 
 export { browserMeta, createSession, sdkMeta } from './metas';
 
@@ -35,7 +35,6 @@ export {
   BaseExtension,
   BaseInstrumentation,
   BaseTransport,
-  Conventions,
   createInternalLogger,
   createPromiseBuffer,
   deepEqual,
@@ -93,12 +92,7 @@ export {
   EVENT_SESSION_RESUME,
   EVENT_SESSION_START,
   EVENT_VIEW_CHANGED,
-  apiMessageBus,
   Observable,
-  USER_ACTION_CANCEL,
-  USER_ACTION_END,
-  USER_ACTION_HALT,
-  USER_ACTION_START,
   unknownString,
 } from '@grafana/faro-core';
 
@@ -163,18 +157,12 @@ export type {
   TransportItemPayload,
   Transports,
   UnpatchedConsole,
-  ApiMessageBusMessages,
-  UserActionStartMessage,
-  UserActionEndMessage,
-  UserActionCancelMessage,
-  UserActionHaltMessage,
 } from '@grafana/faro-core';
 
 export {
   PersistentSessionsManager,
   VolatileSessionsManager,
   MAX_SESSION_PERSISTENCE_TIME,
-  MAX_SESSION_PERSISTENCE_TIME_BUFFER,
   SESSION_EXPIRATION_TIME,
   SESSION_INACTIVITY_TIME,
   STORAGE_KEY,
@@ -182,4 +170,4 @@ export {
 
 export { getIgnoreUrls, getUrlFromResource } from './utils/url';
 
-export { userActionDataAttribute, startUserAction } from './instrumentations/userActions';
+export { userActionDataAttribute } from './instrumentations/userActions';