+++ packages/web-sdk/src/instrumentations/userActions/const.ts
@@ -5,5 +5,3 @@ export const MESSAGE_TYPE_DOM_MUTATION = 'dom-mutation';
 
 export const userActionDataAttributeParsed = 'faroUserActionName';
 export const userActionDataAttribute = 'data-faro-user-action-name';
-
-export const userActionStartByApiCallEventName = 'faroApiCall';