+++ packages/web-tracing/src/index.ts
@@ -1,7 +1,5 @@
 export { FaroTraceExporter } from './faroTraceExporter';
 
-export { FaroSessionSpanProcessor } from './sessionSpanProcessor';
-
 export { getDefaultOTELInstrumentations } from './getDefaultOTELInstrumentations';
 
 export { TracingInstrumentation } from './instrumentation';